+++ examples/backup-example/main.tf
@@ -2,7 +2,10 @@ terraform {
   required_version = "~>0.12.15"
 
   required_providers {
-    azurerm = "~> 1.40"
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = ">= 2.43"
+    }
   }
 }