+++ modules/automation-account/versions.tf
@@ -1,6 +1,9 @@
 terraform {
   required_version = ">= 0.12"
   required_providers {
-    azurerm = ">= 1.32"
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = ">= 2.43"
+    }
   }
 }