+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Changed
+  * AZ-445: Remove datasource to get log_analytics_workspace_id and get it from variable (https://github.com/terraform-providers/terraform-provider-azurerm/pull/10162)
+  * AZ-445: Fix deprecation warns on automation-account
+
 # v3.1.1/v4.0.0 - 2020-11-19
 
 Updated